rename `HiveBridgeHelper` to `KyuubiHiveConnectorBridgeHelper`
